Add files via upload
The process of analyzing language to identify attitudes and opinions, depending on their polarity, is known as sentiment analysis.I built a model in Python for this data science project so that it can analyze tweet sentiment, including positive, negative, and neutral sentiment. The dataset makes the assumption that any tweets with emoticons, such as :), were good, and that tweets with emoticons, such as :(, were negative.